fix bugs in linkage for 64-bit

+++ include/mem.h
@@ -4,9 +4,9 @@
   #include <stdio.h>
   #include "asmdiocommon.h"
 
-  void* PTR32 MALLOC24(size_t len);
-  int FREE24(void* PTR32 addr, size_t len);
-  void* PTR32 MALLOC31(size_t len);
+  void* PTR32 MALLOC24(unsigned int len);
+  int FREE24(void* PTR32 addr, unsigned int len);
+  void* PTR32 MALLOC31(unsigned int len);
   void FREE31(void* PTR32 addr);
 
   void dumpstg(FILE* stream, void* p, size_t len);
